Corrigindo relacionamento appointmentService

+++ src/main/java/com/lasystems/lagenda/dtos/ServiceSummary.java
@@ -1,11 +1,12 @@
 package com.lasystems.lagenda.dtos;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ServiceSummary(
         UUID id,
         String name,
-        Double price,
+        BigDecimal price,
         Integer durationMinutes
 ) {
 }
